on scale re-render only visible pdf pages
