+++ src/components/document/pdf-viewer/Document.js
@@ -1,12 +1,11 @@
 import React, { createRef, useEffect, useRef, useState } from 'react'
 import cn from 'classnames'
-import { Grid, IconButton, Menu, MenuItem, TextField, Toolbar, Tooltip, Typography } from '@material-ui/core'
-import { ArrowDownward, ArrowDropDown, ArrowUpward, Fullscreen, FullscreenExit, ZoomIn, ZoomOut } from '@material-ui/icons'
+import { Typography } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import { STATUS_COMPLETE, STATUS_ERROR, STATUS_LOADING, useDocument } from './DocumentProvider'
-import Page from './Page'
-import { zoomIn, zoomOut } from './zooming'
 import Loading from '../../Loading'
+import Page from './Page'
+import Toolbar from './Toolbar'
+import { STATUS_COMPLETE, STATUS_ERROR, STATUS_LOADING, useDocument } from './DocumentProvider'
 
 const useStyles = makeStyles(theme => ({
     error: {
@@ -46,48 +45,6 @@ const useStyles = makeStyles(theme => ({
             }
         }
     },
-    toolbar: {
-        backgroundColor: theme.palette.grey[100],
-        borderColor: theme.palette.grey[400],
-        borderWidth: 1,
-        borderTopStyle: 'solid',
-        borderBottomStyle: 'solid',
-        justifyContent: 'space-between',
-    },
-    toolbarIcon: {
-        marginRight: theme.spacing(1),
-    },
-    pageInfo: {
-        display: 'inline-flex',
-        alignItems: 'center',
-        '& span': {
-            marginLeft: theme.spacing(0.5)
-        }
-    },
-    pageNumber: {
-        width: 60,
-        backgroundColor: theme.palette.background.default,
-        '& .MuiOutlinedInput-inputMarginDense': {
-            textAlign: 'right',
-            padding: '5px 8px',
-        }
-    },
-    scaleSelect: {
-        width: 120,
-        paddingRight: 4,
-        cursor: 'pointer',
-        border: '1px solid rgba(0, 0, 0, 0.23)',
-        backgroundColor: theme.palette.background.default,
-        '&:hover': {
-            border: '1px solid rgba(0, 0, 0, 0.87)',
-        },
-        '& .MuiOutlinedInput-input': {
-            width: 80,
-            textAlign: 'right',
-            display: 'inline-block',
-            padding: '5px 0 5px 8px',
-        },
-    },
 }))
 
 const pageMargin = 20
@@ -101,7 +58,6 @@ export default function Document({ initialPageIndex, onPageIndexChange, renderer
 
     const viewerRef = useRef()
     const containerRef = useRef()
-    const pageInputRef = useRef()
 
     useEffect(() => {
         if (doc?.numPages) {
@@ -122,7 +78,7 @@ export default function Document({ initialPageIndex, onPageIndexChange, renderer
 
     useEffect(() => {
         if (status === STATUS_COMPLETE && initialPageIndex > 0 && pagesRefs[initialPageIndex]) {
-            scrollToPage(initialPageIndex)
+            containerRef.current.scrollTop = pagesRefs[initialPageIndex].current.offsetTop
         }
     }, [status])
 
@@ -139,168 +95,21 @@ export default function Document({ initialPageIndex, onPageIndexChange, renderer
         }
     }
 
-    const [anchorEl, setAnchorEl] = useState(null)
-    const handleScaleMenuClick = event => setAnchorEl(event.currentTarget)
-    const handleScaleMenuClose = () => setAnchorEl(null)
-    const handleScaleSet = scale => () => {
-        handleScaleMenuClose()
-        const containerWidth = containerRef.current.clientWidth - pageMargin
-        const containerHeight = containerRef.current.clientHeight - pageMargin
-        if (scale === 'page') {
-            setScale(Math.min(containerWidth / firstPageData.width, containerHeight / firstPageData.height))
-        } else if (scale === 'width') {
-            setScale(containerWidth / firstPageData.width)
-        } else {
-            setScale(scale)
-        }
-    }
-
-    const scrollToPage = index => containerRef.current.scrollTop = pagesRefs[index].current.offsetTop
-
-    const onPrevPage = () => scrollToPage(initialPageIndex - 1)
-    const onNextPage = () => scrollToPage(initialPageIndex + 1)
-    const onPageFocus = () => pageInputRef.current.select()
-    const onPageChange = event => {
-        const page = parseInt(event.target.value)
-        if (!isNaN(page) && page > 0 && page <= doc?.numPages) {
-            scrollToPage(page - 1)
-        }
-    }
-
-    const onZoomOut = () => setScale(zoomOut(scale))
-    const onZoomIn = () => setScale(zoomIn(scale))
-
-    const onFullScreen = () => viewerRef.current.requestFullscreen()
-    const onFullScreenExit = () => document.exitFullscreen()
-
-    const popperProps = {
-        container: viewerRef.current
-    }
-
     return (
         <div ref={viewerRef} className={classes.viewer}>
-            <Toolbar variant="dense" classes={{root: classes.toolbar}}>
-                <Grid container justify="space-between">
-                    <Grid item>
-                        <Tooltip title="Previous page" PopperProps={popperProps}>
-                            <span>
-                                <IconButton
-                                    size="small"
-                                    onClick={onPrevPage}
-                                    disabled={status === STATUS_LOADING || initialPageIndex === 0}
-                                    className={classes.toolbarIcon}
-                                >
-                                    <ArrowUpward />
-                                </IconButton>
-                            </span>
-                        </Tooltip>
-                        <Tooltip title="Next page" PopperProps={popperProps}>
-                            <span>
-                                <IconButton
-                                    size="small"
-                                    onClick={onNextPage}
-                                    disabled={status === STATUS_LOADING || initialPageIndex === doc?.numPages - 1}
-                                    className={classes.toolbarIcon}
-                                >
-                                    <ArrowDownward />
-                                </IconButton>
-                            </span>
-                        </Tooltip>
-                        <div className={classes.pageInfo}>
-                            <Tooltip title="Page" PopperProps={popperProps}>
-                                <span>
-                                    <TextField
-                                        size="small"
-                                        variant="outlined"
-                                        inputRef={pageInputRef}
-                                        value={initialPageIndex + 1}
-                                        className={classes.pageNumber}
-                                        onFocus={onPageFocus}
-                                        onChange={onPageChange}
-                                        disabled={status === STATUS_LOADING}
-                                    />
-                                </span>
-                            </Tooltip>
-                            <span>
-                                of
-                            </span>
-                            <span>
-                                {doc?.numPages}
-                            </span>
-                        </div>
-                    </Grid>
-                    <Grid item>
-                        <Tooltip title="Zoom out" PopperProps={popperProps}>
-                            <span>
-                                <IconButton
-                                    size="small"
-                                    onClick={onZoomOut}
-                                    className={classes.toolbarIcon}
-                                    disabled={status === STATUS_LOADING}
-                                >
-                                    <ZoomOut />
-                                </IconButton>
-                            </span>
-                        </Tooltip>
-                        <Tooltip title="Zoom in" PopperProps={popperProps}>
-                            <span>
-                                <IconButton
-                                    size="small"
-                                    onClick={onZoomIn}
-                                    className={classes.toolbarIcon}
-                                    disabled={status === STATUS_LOADING}
-                                >
-                                    <ZoomIn />
-                                </IconButton>
-                            </span>
-                        </Tooltip>
-                        <Tooltip title="Scale" PopperProps={popperProps}>
-                            <div className={cn(
-                                    'MuiInputBase-root',
-                                    'MuiOutlinedInput-root',
-                                    'MuiInputBase-formControl',
-                                    classes.scaleSelect,
-                                    {
-                                        'Mui-disabled': status === STATUS_LOADING
-                                    }
-                                )}
-                                onClick={status === STATUS_COMPLETE ? handleScaleMenuClick : null}
-                            >
-                                <span className={cn(
-                                    'MuiInputBase-input',
-                                    'MuiOutlinedInput-input',
-                                )}>
-                                    {Math.round(scale * 100) + '%'}
-                                </span>
-                                <ArrowDropDown className={cn(
-                                    'MuiSelect-icon',
-                                    'MuiSelect-iconOutlined'
-                                )}/>
-                            </div>
-                        </Tooltip>
-                    </Grid>
-                    <Grid item>
-                        <Tooltip title="Full screen" PopperProps={popperProps}>
-                            <IconButton
-                                size="small"
-                                onClick={onFullScreen}
-                                className={classes.fullscreen}
-                            >
-                                <Fullscreen />
-                            </IconButton>
-                        </Tooltip>
-                        <Tooltip title="Exit full screen" PopperProps={popperProps}>
-                            <IconButton
-                                size="small"
-                                onClick={onFullScreenExit}
-                                className={classes.fullscreenExit}
-                            >
-                                <FullscreenExit />
-                            </IconButton>
-                        </Tooltip>
-                    </Grid>
-                </Grid>
-            </Toolbar>
+            <Toolbar
+                viewerRef={viewerRef}
+                containerRef={containerRef}
+                pagesRefs={pagesRefs}
+                initialPageIndex={initialPageIndex}
+                numPages={doc?.numPages}
+                firstPageData={firstPageData}
+                pageMargin={pageMargin}
+                scale={scale}
+                setScale={setScale}
+                fullscreenClass={classes.fullscreen}
+                fullscreenExitClass={classes.fullscreenExit}
+            />
 
             <div className={cn(classes.container, 'pdfViewer')} ref={containerRef}>
                 {status === STATUS_LOADING && (
@@ -333,26 +142,6 @@ export default function Document({ initialPageIndex, onPageIndexChange, renderer
                     })
                 }
             </div>
-
-            <Menu
-                anchorEl={anchorEl}
-                open={Boolean(anchorEl)}
-                onClose={handleScaleMenuClose}
-                disableScrollLock={true}
-                {...popperProps}
-            >
-                <MenuItem onClick={handleScaleSet(1)}>Original</MenuItem>
-                <MenuItem onClick={handleScaleSet('page')}>Page fit</MenuItem>
-                <MenuItem onClick={handleScaleSet('width')}>Page width</MenuItem>
-                <MenuItem onClick={handleScaleSet(0.5)}>50%</MenuItem>
-                <MenuItem onClick={handleScaleSet(0.75)}>75%</MenuItem>
-                <MenuItem onClick={handleScaleSet(1)}>100%</MenuItem>
-                <MenuItem onClick={handleScaleSet(1.25)}>125%</MenuItem>
-                <MenuItem onClick={handleScaleSet(1.5)}>150%</MenuItem>
-                <MenuItem onClick={handleScaleSet(2)}>200%</MenuItem>
-                <MenuItem onClick={handleScaleSet(3)}>300%</MenuItem>
-                <MenuItem onClick={handleScaleSet(4)}>400%</MenuItem>
-            </Menu>
         </div>
     )
 }
